add hook for save in ordercontactgrid

+++ src/RentalWorksWeb/Libraries/Fw/scripts/FwControls/FwBrowse.ts
@@ -1810,7 +1810,11 @@ class FwBrowseClass {
                                                 $confirmation.find('.recordno').html((i + 1).toString());
                                                 $confirmation.find('.progress').attr('value', (i + 1).toString());
                                                 let $tr = $selectedCheckBoxes.eq(i).closest('tr');
-                                                await this.deleteRecord(options.$browse, $tr);
+                                                if (typeof options.$browse.data('deletewithnoids') === 'function') {
+                                                    options.$browse.data('deletewithnoids')(options.$browse, $tr);
+                                                } else {
+                                                    await this.deleteRecord(options.$browse, $tr);
+                                                }
                                             }
                                         } catch (ex) {
                                             FwFunc.showError(ex);
@@ -3827,8 +3831,12 @@ class FwBrowseClass {
 
             $ok.on('click', async () => {
                 try {
-                    await me.deleteRecord($control, $tr);
-                    await me.databind($control);
+                    if (typeof $control.data('deletewithnoids') === 'function') {
+                        $control.data('deletewithnoids')($control, $tr);
+                    } else {
+                        await me.deleteRecord($control, $tr);
+                        await me.databind($control);
+                    }
                 } catch (ex) {
                     FwFunc.showError(ex);
                 }
@@ -3837,80 +3845,80 @@ class FwBrowseClass {
     }
     //----------------------------------------------------------------------------------------------
     async deleteRecord($control: JQuery, $tr: JQuery): Promise<void> {
-        return new Promise<void>(async (resolve, reject) => {
-            try {
-                let me = this;
-                this.autoSave($control, $tr);
-                let miscfields = {};
-                let name = $control.attr('data-name');
-                let $form = $control.closest('.fwform');
-                let rowuniqueids = this.getRowFormUniqueIds($control, $tr);
-                const request = new FwAjaxRequest<any>();
-                request.data = {
-                    module: name,
-                    ids: rowuniqueids,
-                    miscfields: miscfields
-                };
-                if ($form.length > 0) {
-                    let formuniqueids = ($form.length > 0) ? FwModule.getFormUniqueIds($form) : [];
-                    request.data.miscfields = jQuery.extend({}, miscfields, formuniqueids);
-                }
-
-                var controller: any = window[name + 'Controller'];
-                if (typeof controller === 'undefined') {
-                    throw name + 'Controller is not defined.'
-                }
-                let url = '';
-                if (typeof $control.data('getapiurl') === 'function') {
-                    url = $control.data('getapiurl')('DELETE');
-                }
-                if (url.length === 0 && typeof controller.apiurl !== 'undefined' || typeof $control.data('getbaseapiurl') === 'function') {
-                    if (typeof $control.data('getbaseapiurl') !== 'undefined') {
-                        url = $control.data('getbaseapiurl')();
+            return new Promise<void>(async (resolve, reject) => {
+                try {
+                    let me = this;
+                    this.autoSave($control, $tr);
+                    let miscfields = {};
+                    let name = $control.attr('data-name');
+                    let $form = $control.closest('.fwform');
+                    let rowuniqueids = this.getRowFormUniqueIds($control, $tr);
+                    const request = new FwAjaxRequest<any>();
+                    request.data = {
+                        module: name,
+                        ids: rowuniqueids,
+                        miscfields: miscfields
+                    };
+                    if ($form.length > 0) {
+                        let formuniqueids = ($form.length > 0) ? FwModule.getFormUniqueIds($form) : [];
+                        request.data.miscfields = jQuery.extend({}, miscfields, formuniqueids);
                     }
-                    else if (typeof controller.apiurl !== 'undefined' && controller.apiurl.length > 0) {
-                        url = controller.apiurl;
+
+                    var controller: any = window[name + 'Controller'];
+                    if (typeof controller === 'undefined') {
+                        throw name + 'Controller is not defined.'
+                    }
+                    let url = '';
+                    if (typeof $control.data('getapiurl') === 'function') {
+                        url = $control.data('getapiurl')('DELETE');
+                    }
+                    if (url.length === 0 && typeof controller.apiurl !== 'undefined' || typeof $control.data('getbaseapiurl') === 'function') {
+                        if (typeof $control.data('getbaseapiurl') !== 'undefined') {
+                            url = $control.data('getbaseapiurl')();
+                        }
+                        else if (typeof controller.apiurl !== 'undefined' && controller.apiurl.length > 0) {
+                            url = controller.apiurl;
+                        }
+                        else {
+                            throw `No apiurl defined for Grid: ${name}`;
+                        }
+                        var ids: any = [];
+                        for (var key in request.data.ids) {
+                            ids.push(request.data.ids[key].value);
+                        }
+                        ids = ids.join('~');
+                        if (ids.length === 0) {
+                            throw 'primary key id(s) cannot be blank';
+                        }
+                        url += '/' + ids;
+                    }
+                    request.url = applicationConfig.apiurl + url;
+                    request.httpMethod = 'DELETE';
+
+                    const response = await FwAjax.callWebApi<any, any>(request)
+                    if (request.xmlHttpRequest.status === 200 || request.xmlHttpRequest.status === 404) {
+                        //perform after delete
+                        if (($control.attr('data-type') === 'Grid') && (typeof $control.data('afterdelete') === 'function')) {
+                            $control.data('afterdelete')($control, $tr);
+                        }
+                        else if (($control.attr('data-type') == 'Grid') && (typeof $control.attr('data-controller') !== 'undefined') && ($control.attr('data-controller') !== '')) {
+                            if (controller.afterDelete === 'function') {
+                                controller.afterDelete($control, $tr);
+                            }
+                        }
+                        //if (refreshAfterDelete) {
+                        //    me.search($control);
+                        //}
+                        resolve();
                     }
                     else {
-                        throw `No apiurl defined for Grid: ${name}`;
+                        reject(response);
                     }
-                    var ids: any = [];
-                    for (var key in request.data.ids) {
-                        ids.push(request.data.ids[key].value);
-                    }
-                    ids = ids.join('~');
-                    if (ids.length === 0) {
-                        throw 'primary key id(s) cannot be blank';
-                    }
-                    url += '/' + ids;
                 }
-                request.url = applicationConfig.apiurl + url;
-                request.httpMethod = 'DELETE';
-
-                const response = await FwAjax.callWebApi<any, any>(request)
-                if (request.xmlHttpRequest.status === 200 || request.xmlHttpRequest.status === 404) {
-                    //perform after delete
-                    if (($control.attr('data-type') === 'Grid') && (typeof $control.data('afterdelete') === 'function')) {
-                        $control.data('afterdelete')($control, $tr);
-                    }
-                    else if (($control.attr('data-type') == 'Grid') && (typeof $control.attr('data-controller') !== 'undefined') && ($control.attr('data-controller') !== '')) {
-                        if (controller.afterDelete === 'function') {
-                            controller.afterDelete($control, $tr);
-                        }
-                    }
-                    //if (refreshAfterDelete) {
-                    //    me.search($control);
-                    //}
-                    resolve();
+                catch (ex) {
+                    reject(ex);
                 }
-                else {
-                    reject(response);
-                }
-            }
-            catch (ex) {
-                reject(ex);
-            }
-        });
+            });
     }
     //---------------------------------------------------------------------------------
     addLegend($control: JQuery, caption: string, color: string) {
